% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated data-set.}
\format{A data frame with 4000 rows and 5 variables:
\describe{
  \item{c_id}{cluster id}
  \item{s_id}{subject id}
  \item{t}{time}
  \item{cov}{covariate}
  \item{y}{outcome}
  ...
}}
\usage{
sim_data
}
\description{
A dataset containing three-level ordinal outcomes
for 250 clusters, 1000 subjects, 4 time points
}
\keyword{datasets}
